Move ReactContext source to React package (#12205)
